JAVA-22042 | Removed cargo plugin in favor or mockserver. removed duplicated mockserver implementation
