+++ apache-httpclient4/src/test/java/com/baeldung/httpclient/HttpAsyncClientV4LiveTest.java
@@ -31,6 +31,7 @@ import org.apache.http.protocol.HttpContext;
 import org.apache.http.ssl.SSLContexts;
 import org.junit.jupiter.api.Test;
 
+import com.baeldung.GetRequestMockServer;
 
 class HttpAsyncClientV4LiveTest extends GetRequestMockServer {
