+++ apache-httpclient4/src/test/java/com/baeldung/httpclient/httpclient/ApacheHttpClientUnitTest.java
@@ -13,13 +13,15 @@ import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.util.EntityUtils;
 import org.junit.jupiter.api.Test;
 
+import com.baeldung.GetRequestMockServer;
+
 class ApacheHttpClientUnitTest extends GetRequestMockServer {
 
 
     @Test
     void givenDeveloperUsedCloseableHttpResponse_whenExecutingGetRequest_thenStatusIsOk() throws IOException {
         try (CloseableHttpClient httpClient = HttpClientBuilder.create().build()) {
-            HttpGet httpGet = new HttpGet(serviceOneUrl);
+            HttpGet httpGet = new HttpGet(simplePathUrl);
             try (CloseableHttpResponse response = httpClient.execute(httpGet)) {
                 HttpEntity entity = response.getEntity();
                 EntityUtils.consume(entity);
